docs: add readme
