+++ packages/graphin/src/hooks/useGraph.ts
@@ -3,11 +3,12 @@ import { Graph } from '@antv/g6';
 import type { GraphinProps } from '../types';
 
 /**
- *
- * @param props
+ * Hook for creating and managing a G6 graph instance.
+ * @param props The props for the Graphin component.
+ * @returns An object containing the graph instance, the container ref, and a boolean indicating whether the graph is ready.
  */
 export default function useGraph<P extends GraphinProps>(props: P) {
-  const { onInit, onRender, onDestroy, options } = props;
+  const { onInit, onReady, onDestroy, options } = props;
   const [isReady, setIsReady] = useState(false);
   const graphRef: React.MutableRefObject<Graph | null> = useRef(null);
   const containerRef: React.MutableRefObject<HTMLDivElement | null> = useRef(null);
@@ -25,7 +26,7 @@ export default function useGraph<P extends GraphinProps>(props: P) {
       const graph = graphRef.current;
       if (graph) {
         graph.destroy();
-        onDestroy?.(graph);
+        onDestroy?.();
         graphRef.current = null;
       }
     };
@@ -38,7 +39,7 @@ export default function useGraph<P extends GraphinProps>(props: P) {
     if (!options || !container || !graph || graph.destroyed) return;
 
     graph.setOptions(options);
-    graph.render().then(() => onRender?.(graph));
+    graph.render().then(() => onReady?.(graph));
   }, [options]);
 
   return {
