+++ packages/graphin/src/index.ts
@@ -1,5 +1,3 @@
-import Graphin from './Graphin';
-
+export { Graphin } from './Graphin';
 export { GraphinContext, useGraphin } from './context';
 export * from './types';
-export default Graphin;
